Fix #56. Now the namespace csa is used.
XPath seems to require a prefix. The introduced prefix is csa. All types
are defined in that namespace. In the map file now one has to use
`csa:uamapping` instead of `uamapping`. All nodes inside `csa:uamapping`
can keep the original name, because the default element form is set to
`unqualified` in the xml scheme.