Configuration option for encrypted communication (#51)
* feat(server) include endpoint security configuration

* feat(adapter) parsing of certificate folder, encryption configuration added

* feat(adapter) add certificate folder structure

* Format file. Fix blockListFolder and issuerListfolder typo.

* Throw exception in case setting up the endpoints fails.

* Use port from mapping file.

* Separate filling of the build info. Fill it after endpoint creation!

* Add security configuration to the mapping example.

* Fix documentation.

* Update xml schema by adding security information.

* Fix #56. Now the namespace csa is used.

XPath seems to require a prefix. The introduced prefix is csa. All types
are defined in that namespace. In the map file now one has to use
`csa:uamapping` instead of `uamapping`. All nodes inside `csa:uamapping`
can keep the original name, because the default element form is set to
`unqualified` in the xml scheme.

* Use standard application URI.

At least UAExpert was complaining about that application URI when using
a certificate that used the previous application URI opc.tcp://HOSTNAME.
Now the URI is set to urn:HOSTNAME:ChimeraTK:APPNAME

Co-authored-by: Andreas Ebner <ebner.andreas@outlook.com>
Co-authored-by: Zenker, Dr. Klaus (FWKE) - 126506 <k.zenker@hzdr.de>