fix: Create void event type before using it. Reduced log output.
The void event type was created in the worker thread and used in the
void observer thread. It could happen that the first void event was
fired before the void event type was created.

Reduced output by checking if process variables are used in histories.
In that case do not warn about not found source variables, because the
histories are created with respect to the mapped variable tree. Also do
not print all history pvs and mapped pvs when using logging level info,
but only when using logging level debug.